feat: update README with project state
It's probably a good idea to inform people looking at the README
that the project is still 100% unusable. That way, if someone is
looking for an actual ZMTP implementation, they can move on in
their search.

Also, I added the MPL license header to the README. Not sure if
that's necessary in a README, but might as well. Bytes are cheap.